+++ postgres/mod.ts
@@ -1,4 +1,3 @@
 export * from "./postgres-migrator.ts";
-// export * from "./postgres-service.ts";
-// export * from "./postgres-store.ts";
 export * from "./postgres-client.ts";
+export * from "./tables.ts";